Create A1 Project Skeleton
PolygonalMind uses child collider meshes, except they import as mesh renderers rather than as colliders. There might be a way to automate this in the future. But for now, I swapped them manually. The environment seems like a good test for character controllers, given it is designed for player avatars to explore.