Start Lab-C3 and add ComponentBroker
Hack Lab - C3 is intended to be a test playground for various Unika operations. However, I haven't actually added any test code yet.

Instead, I've added the ComponentBroker, which is a crazy design that will be interesting to see how well it performs. And while I was at it, I decided to add a PhysicsAspectLookup since I don't think Unity is going to change that interface any time soon. It isn't the most pretty, but it is at least functional. Unfortunately, it doesn't seem possible to do a similar thing with ComponentBroker. But at least for ComponentBroker, extension methods are viable.

My plan A for authoring won't work, since the authoring type cannot have direct access to a script's private members. So next up, enumerator filters.