Make Unika spin a cube
Unika can do it now!

Wow, was there a lot that needed to happen to reach this point. Stuff with ComponentBroker, stuff with aspects, stuff with the source generators, stuff with enumerators, and a nasty bug involving default interface methods that needed to be overcome.

I still need to test entity serialization, interface return types, and a few other loose ends, not to mention authoring. So there's still plenty to do!