Add ScriptRef test and fix bugs
There was a baking bug where I accidentally replicated the first script data for all the other scripts. And I also had an infinite recursion bug using the wrong overload. Plus, there were a couple of source-generator bugs to fix too.

But with that all said and done, I'm surprised my tests are working the way they are. Time will tell if this tech is really as stable as it seems right now (probably not).