Cleanup and fix LifeFX
I'm dropping LifeFX sync states for now. I think partial struct uploads will be vital for performance, and my design didn't support that. That also makes the compute shader very complicated, and not something that can be generated from a template. So I need to rethink that design.

Also, ScriptableObjects get duplicated in a build when they are in both a scene and a subscene. So I have added an asset hash remap mechanism. I managed to do this in a way that avoids ECS data holding the hashes, so I'm fairly happy about that.