Make Burst and Unika work
Kinda a mess, because there's also a big sync with LSSS in here, but with Burst 1.8.19, I managed to get something to work. It is not great though. I had to kill some of the API, and migrate other API into source gen. But at least my test example works.